Invert api/praxis.ts and api/duel.ts's upward import of castTallies (#2893)
Both cleared cast tallies through components/vote/castTallies.ts, an api
module importing a component. The store (recordCastTally,
clearCastTallies, castTally, __resetCastTallies) has no React in it and
moves to the new utils/castTallies.ts; components/vote/castTallies.ts
keeps only useCastTally, the useSyncExternalStore hook wrapper, which now
reads the store from utils/. Test files that reached the seam through the
old path are updated to import from utils/castTallies.ts directly, so the
#2697 test-seam ratchet still finds a single declaring owner.
