+++ frontend/src/components/praxisCard/scoreStamp/__tests__/scoreStamp.test.tsx
@@ -21,7 +21,7 @@ import { surfaceMap } from '../../../../factions'
 import { scoreBreakdown, formatMult } from '../scoreBreakdown'
 import { formatPoints } from '../../../../utils/points'
 import { applyCastTally } from '../../../vote/useVotedPraxis'
-import { recordCastTally, castTally, __resetCastTallies } from '../../../vote/castTallies'
+import { recordCastTally, castTally, __resetCastTallies } from '../../../../utils/castTallies'
 import ScoreStamp from '../ScoreStamp'
 import DefaultScoreStamp from '../DefaultScoreStamp'
 import EverymenScoreStamp from '../EverymenScoreStamp'
